[iOS] refactor: TabBarCoordinator에서 ViewController 생성 분리
